requirements.txt
# Database
sqlalchemy>=1.4.0
pymysql>=1.0.0
alembic>=1.7.0

# Environment and configuration
python-dotenv>=0.19.0

# Web and API
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
starlette>=0.26.0

# Google Calendar integration
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0

# Utilities
requests>=2.25.0
pydantic>=1.9.0
email-validator>=1.1.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Development
black>=21.5b2
flake8>=3.9.0